Refactor and restructure Paper package organization
