+++ paper/src/main/java/net/thenextlvl/commander/paper/CommanderPlugin.java
@@ -1,4 +1,4 @@
-package net.thenextlvl.commander;
+package net.thenextlvl.commander.paper;
 
 import core.i18n.file.ComponentBundle;
 import lombok.Getter;
@@ -6,12 +6,11 @@ import lombok.experimental.Accessors;
 import net.kyori.adventure.text.minimessage.MiniMessage;
 import net.kyori.adventure.text.minimessage.tag.resolver.Placeholder;
 import net.kyori.adventure.text.minimessage.tag.resolver.TagResolver;
-import net.thenextlvl.commander.api.CommandRegistry;
-import net.thenextlvl.commander.api.Commander;
-import net.thenextlvl.commander.command.CommanderCommand;
-import net.thenextlvl.commander.implementation.PaperCommandRegistry;
-import net.thenextlvl.commander.implementation.PaperPermissionOverride;
-import net.thenextlvl.commander.listener.CommandListener;
+import net.thenextlvl.commander.Commander;
+import net.thenextlvl.commander.paper.command.CommanderCommand;
+import net.thenextlvl.commander.paper.implementation.PaperCommandRegistry;
+import net.thenextlvl.commander.paper.implementation.PaperPermissionOverride;
+import net.thenextlvl.commander.paper.listener.CommandListener;
 import org.bstats.bukkit.Metrics;
 import org.bukkit.Bukkit;
 import org.bukkit.entity.Player;
@@ -36,9 +35,8 @@ public class CommanderPlugin extends JavaPlugin implements Commander {
                     Placeholder.component("prefix", bundle.component(Locale.US, "prefix"))
             )).build());
 
-    private final CommandRegistry commandRegistry = new PaperCommandRegistry();
-    private final PaperCommandRegistry commandManager = new PaperCommandRegistry();
-    private final PaperPermissionOverride permissionOverride = new PaperPermissionOverride();
+    private final PaperCommandRegistry commandRegistry = new PaperCommandRegistry(this);
+    private final PaperPermissionOverride permissionOverride = new PaperPermissionOverride(this);
 
     @Override
     public void onLoad() {
@@ -47,13 +45,19 @@ public class CommanderPlugin extends JavaPlugin implements Commander {
 
     @Override
     public void onEnable() {
-        Bukkit.getGlobalRegionScheduler().execute(this, () -> permissionOverride().overridePermissions());
+        Bukkit.getGlobalRegionScheduler().execute(this, () -> {
+            commandRegistry().unregisterCommands();
+            permissionOverride().overridePermissions();
+        });
         registerListeners();
         registerCommands();
     }
 
     @Override
     public void onDisable() {
+        commandRegistry().getHiddenFile().save();
+        commandRegistry().getUnregisteredFile().save();
+        permissionOverride().getOverridesFile().save();
         metrics.shutdown();
     }
