+++ paper/src/main/java/net/thenextlvl/commander/paper/command/HideCommand.java
@@ -1,4 +1,4 @@
-package net.thenextlvl.commander.command;
+package net.thenextlvl.commander.paper.command;
 
 import com.mojang.brigadier.Command;
 import com.mojang.brigadier.arguments.StringArgumentType;
@@ -8,38 +8,37 @@ import io.papermc.paper.command.brigadier.CommandSourceStack;
 import io.papermc.paper.command.brigadier.Commands;
 import lombok.RequiredArgsConstructor;
 import net.kyori.adventure.text.minimessage.tag.resolver.Placeholder;
-import net.thenextlvl.commander.CommanderPlugin;
+import net.thenextlvl.commander.paper.CommanderPlugin;
+import org.bukkit.Bukkit;
+import org.bukkit.entity.Player;
 
 @RequiredArgsConstructor
 @SuppressWarnings("UnstableApiUsage")
-class ResetCommand {
+class HideCommand {
     private final CommanderPlugin plugin;
 
-    public ArgumentBuilder<CommandSourceStack, ?> create() {
-        return Commands.literal("reset")
+    ArgumentBuilder<CommandSourceStack, ?> create() {
+        return Commands.literal("hide")
                 .then(Commands.argument("command", StringArgumentType.string())
                         .suggests((context, suggestions) -> {
-                            plugin.commandRegistry().hiddenCommands().stream()
-                                    .filter(s -> s.contains(suggestions.getRemaining()))
+                            Bukkit.getCommandMap().getKnownCommands().values().stream()
+                                    .map(org.bukkit.command.Command::getLabel)
+                                    .filter(s -> !plugin.commandRegistry().isHidden(s))
                                     .map(StringArgumentType::escapeIfRequired)
-                                    .forEach(suggestions::suggest);
-                            plugin.commandRegistry().unregisteredCommands().stream()
                                     .filter(s -> s.contains(suggestions.getRemaining()))
-                                    .map(StringArgumentType::escapeIfRequired)
                                     .forEach(suggestions::suggest);
                             return suggestions.buildFuture();
                         })
-                        .executes(this::reset));
+                        .executes(this::hide));
     }
 
-    private int reset(CommandContext<CommandSourceStack> context) {
+    private int hide(CommandContext<CommandSourceStack> context) {
         var sender = context.getSource().getSender();
         var command = context.getArgument("command", String.class);
-        var s1 = plugin.permissionOverride().reset(command);
-        var s3 = plugin.commandRegistry().register(command);
-        var s2 = plugin.commandRegistry().reveal(command);
-        var message = s1 || s2 || s3 ? "command.reset" : "nothing.changed";
+        var success = plugin.commandRegistry().hide(command);
+        var message = success ? "command.hidden" : "nothing.changed";
         plugin.bundle().sendMessage(sender, message, Placeholder.parsed("command", command));
+        if (success) Bukkit.getOnlinePlayers().forEach(Player::updateCommands);
         return Command.SINGLE_SUCCESS;
     }
 }
