+++ paper/src/main/java/net/thenextlvl/commander/paper/command/PermissionQueryCommand.java
@@ -1,4 +1,4 @@
-package net.thenextlvl.commander.command;
+package net.thenextlvl.commander.paper.command;
 
 import com.mojang.brigadier.arguments.StringArgumentType;
 import com.mojang.brigadier.builder.ArgumentBuilder;
@@ -7,38 +7,40 @@ import io.papermc.paper.command.brigadier.CommandSourceStack;
 import io.papermc.paper.command.brigadier.Commands;
 import lombok.RequiredArgsConstructor;
 import net.kyori.adventure.text.minimessage.tag.resolver.Placeholder;
-import net.thenextlvl.commander.CommanderPlugin;
+import net.thenextlvl.commander.paper.CommanderPlugin;
 import org.bukkit.Bukkit;
 import org.bukkit.command.Command;
-import org.bukkit.entity.Player;
 
 @RequiredArgsConstructor
 @SuppressWarnings("UnstableApiUsage")
-class UnregisterCommand {
+class PermissionQueryCommand {
     private final CommanderPlugin plugin;
 
     public ArgumentBuilder<CommandSourceStack, ?> create() {
-        return Commands.literal("unregister")
+        return Commands.literal("query")
                 .then(Commands.argument("command", StringArgumentType.string())
                         .suggests((context, suggestions) -> {
                             Bukkit.getCommandMap().getKnownCommands().values().stream()
                                     .map(Command::getLabel)
                                     .filter(s -> !plugin.commandRegistry().isUnregistered(s))
-                                    .filter(s -> s.contains(suggestions.getRemaining()))
                                     .map(StringArgumentType::escapeIfRequired)
+                                    .filter(s -> s.contains(suggestions.getRemaining()))
                                     .forEach(suggestions::suggest);
                             return suggestions.buildFuture();
                         })
-                        .executes(this::unregister));
+                        .executes(this::query));
     }
 
-    private int unregister(CommandContext<CommandSourceStack> context) {
+    private int query(CommandContext<CommandSourceStack> context) {
         var sender = context.getSource().getSender();
         var command = context.getArgument("command", String.class);
-        var success = plugin.commandRegistry().unregister(command);
-        var message = success ? "command.unregistered" : "nothing.changed";
-        plugin.bundle().sendMessage(sender, message, Placeholder.parsed("command", command));
-        if (success) Bukkit.getOnlinePlayers().forEach(Player::updateCommands);
+        var registered = Bukkit.getCommandMap().getKnownCommands().get(command);
+        var permission = registered != null ? registered.getPermission() : null;
+        var message = registered == null ? "command.unknown" : permission != null ?
+                "permission.query.defined" : "permission.query.undefined";
+        plugin.bundle().sendMessage(sender, message,
+                Placeholder.parsed("permission", String.valueOf(permission)),
+                Placeholder.parsed("command", command));
         return com.mojang.brigadier.Command.SINGLE_SUCCESS;
     }
 }
