+++ paper/src/main/java/net/thenextlvl/commander/paper/command/PermissionResetCommand.java
@@ -1,4 +1,4 @@
-package net.thenextlvl.commander.command;
+package net.thenextlvl.commander.paper.command;
 
 import com.mojang.brigadier.Command;
 import com.mojang.brigadier.arguments.StringArgumentType;
@@ -8,34 +8,40 @@ import io.papermc.paper.command.brigadier.CommandSourceStack;
 import io.papermc.paper.command.brigadier.Commands;
 import lombok.RequiredArgsConstructor;
 import net.kyori.adventure.text.minimessage.tag.resolver.Placeholder;
-import net.thenextlvl.commander.CommanderPlugin;
+import net.thenextlvl.commander.paper.CommanderPlugin;
 import org.bukkit.Bukkit;
 import org.bukkit.entity.Player;
 
+import java.util.Optional;
+
 @RequiredArgsConstructor
 @SuppressWarnings("UnstableApiUsage")
-class RegisterCommand {
+class PermissionResetCommand {
     private final CommanderPlugin plugin;
 
     public ArgumentBuilder<CommandSourceStack, ?> create() {
-        return Commands.literal("register")
+        return Commands.literal("reset")
                 .then(Commands.argument("command", StringArgumentType.string())
                         .suggests((context, suggestions) -> {
-                            plugin.commandRegistry().unregisteredCommands().stream()
-                                    .filter(s -> s.contains(suggestions.getRemaining()))
+                            plugin.permissionOverride().originalPermissions().keySet().stream()
                                     .map(StringArgumentType::escapeIfRequired)
+                                    .filter(s -> s.contains(suggestions.getRemaining()))
                                     .forEach(suggestions::suggest);
                             return suggestions.buildFuture();
                         })
-                        .executes(this::register));
+                        .executes(this::reset));
     }
 
-    private int register(CommandContext<CommandSourceStack> context) {
+    private int reset(CommandContext<CommandSourceStack> context) {
         var sender = context.getSource().getSender();
         var command = context.getArgument("command", String.class);
-        var success = plugin.commandRegistry().register(command);
-        var message = success ? "command.registered" : "nothing.changed";
-        plugin.bundle().sendMessage(sender, message, Placeholder.parsed("command", command));
+        var success = plugin.permissionOverride().reset(command);
+        var message = success ? "permission.reset" : "nothing.changed";
+        var permission = Optional.ofNullable(Bukkit.getCommandMap().getCommand(command))
+                .map(org.bukkit.command.Command::getPermission)
+                .orElse("null");
+        plugin.bundle().sendMessage(sender, message, Placeholder.parsed("command", command),
+                Placeholder.parsed("permission", permission));
         if (success) Bukkit.getOnlinePlayers().forEach(Player::updateCommands);
         return Command.SINGLE_SUCCESS;
     }
