+++ paper/src/main/java/net/thenextlvl/commander/paper/command/PermissionSetCommand.java
@@ -1,4 +1,4 @@
-package net.thenextlvl.commander.command;
+package net.thenextlvl.commander.paper.command;
 
 import com.mojang.brigadier.arguments.StringArgumentType;
 import com.mojang.brigadier.builder.ArgumentBuilder;
@@ -7,7 +7,7 @@ import io.papermc.paper.command.brigadier.CommandSourceStack;
 import io.papermc.paper.command.brigadier.Commands;
 import lombok.RequiredArgsConstructor;
 import net.kyori.adventure.text.minimessage.tag.resolver.Placeholder;
-import net.thenextlvl.commander.CommanderPlugin;
+import net.thenextlvl.commander.paper.CommanderPlugin;
 import org.bukkit.Bukkit;
 import org.bukkit.command.Command;
 import org.bukkit.entity.Player;
@@ -24,8 +24,8 @@ class PermissionSetCommand {
                         .suggests((context, suggestions) -> {
                             Bukkit.getCommandMap().getKnownCommands().values().stream()
                                     .map(Command::getLabel)
-                                    .filter(s -> s.contains(suggestions.getRemaining()))
                                     .map(StringArgumentType::escapeIfRequired)
+                                    .filter(s -> s.contains(suggestions.getRemaining()))
                                     .forEach(suggestions::suggest);
                             return suggestions.buildFuture();
                         })
@@ -33,8 +33,8 @@ class PermissionSetCommand {
                                 .suggests((context, suggestions) -> {
                                     Bukkit.getPluginManager().getPermissions().stream()
                                             .map(Permission::getName)
-                                            .filter(s -> s.contains(suggestions.getRemaining()))
                                             .map(StringArgumentType::escapeIfRequired)
+                                            .filter(s -> s.contains(suggestions.getRemaining()))
                                             .forEach(suggestions::suggest);
                                     return suggestions.buildFuture();
                                 })
