+++ paper/src/main/java/net/thenextlvl/commander/paper/command/ResetCommand.java
@@ -1,4 +1,4 @@
-package net.thenextlvl.commander.command;
+package net.thenextlvl.commander.paper.command;
 
 import com.mojang.brigadier.Command;
 import com.mojang.brigadier.arguments.StringArgumentType;
@@ -8,37 +8,38 @@ import io.papermc.paper.command.brigadier.CommandSourceStack;
 import io.papermc.paper.command.brigadier.Commands;
 import lombok.RequiredArgsConstructor;
 import net.kyori.adventure.text.minimessage.tag.resolver.Placeholder;
-import net.thenextlvl.commander.CommanderPlugin;
-import org.bukkit.Bukkit;
-import org.bukkit.entity.Player;
+import net.thenextlvl.commander.paper.CommanderPlugin;
 
 @RequiredArgsConstructor
 @SuppressWarnings("UnstableApiUsage")
-class HideCommand {
+class ResetCommand {
     private final CommanderPlugin plugin;
 
-    ArgumentBuilder<CommandSourceStack, ?> create() {
-        return Commands.literal("hide")
+    public ArgumentBuilder<CommandSourceStack, ?> create() {
+        return Commands.literal("reset")
                 .then(Commands.argument("command", StringArgumentType.string())
                         .suggests((context, suggestions) -> {
-                            Bukkit.getCommandMap().getKnownCommands().values().stream()
-                                    .map(org.bukkit.command.Command::getLabel)
-                                    .filter(s -> !plugin.commandRegistry().isHidden(s))
-                                    .filter(s -> s.contains(suggestions.getRemaining()))
+                            plugin.commandRegistry().hiddenCommands().stream()
                                     .map(StringArgumentType::escapeIfRequired)
+                                    .filter(s -> s.contains(suggestions.getRemaining()))
+                                    .forEach(suggestions::suggest);
+                            plugin.commandRegistry().unregisteredCommands().stream()
+                                    .map(StringArgumentType::escapeIfRequired)
+                                    .filter(s -> s.contains(suggestions.getRemaining()))
                                     .forEach(suggestions::suggest);
                             return suggestions.buildFuture();
                         })
-                        .executes(this::hide));
+                        .executes(this::reset));
     }
 
-    private int hide(CommandContext<CommandSourceStack> context) {
+    private int reset(CommandContext<CommandSourceStack> context) {
         var sender = context.getSource().getSender();
         var command = context.getArgument("command", String.class);
-        var success = plugin.commandRegistry().hide(command);
-        var message = success ? "command.hidden" : "nothing.changed";
+        var s1 = plugin.permissionOverride().reset(command);
+        var s3 = plugin.commandRegistry().register(command);
+        var s2 = plugin.commandRegistry().reveal(command);
+        var message = s1 || s2 || s3 ? "command.reset" : "nothing.changed";
         plugin.bundle().sendMessage(sender, message, Placeholder.parsed("command", command));
-        if (success) Bukkit.getOnlinePlayers().forEach(Player::updateCommands);
         return Command.SINGLE_SUCCESS;
     }
 }
